refactor: collocate TableList and UserList components with tests - Removed legacy AppList and UserList implementations to streamline the codebase.
